Successful Jest tests performed
Added unit tests using Jest for the Bowling class. Verified the correct update of throw and score arrays, detection of strikes and spares, as well as the final bowling score calculation. The tests were executed successfully and the expected results were obtained. The code was modified to enable modularization and facilitate the testing process.

Co-authored-by: DerekAyala <derek.ayala12345@hotmail.com>